# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DietCoach AI Backend API",
    "version": "2.0.0",
    "description": "Meal planning API with AI-powered meal generation and external recipe search"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/preferences": {
      "get": {
        "summary": "Get user preferences",
        "tags": ["Configuration"],
        "responses": {
          "200": {
            "description": "Preferences retrieved"
          }
        }
      },
      "post": {
        "summary": "Create or update preferences",
        "tags": ["Configuration"],
        "responses": {
          "200": {
            "description": "Preferences updated"
          },
          "201": {
            "description": "Preferences created"
          }
        }
      }
    },
...
    "/recipes/{id}": {
      "get": {
        "summary": "Get recipe by ID",
        "tags": ["Recipes"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe retrieved"
          },
          "404": {
            "description": "Recipe not found"
          }
        }
      }
    },
...
    "/meal-plans/{id}": {
      "get": {
        "summary": "Get meal plan with details",
        "tags": ["Meal Plans"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meal plan retrieved"
          },
          "404": {
            "description": "Meal plan not found"
          }
        }
      }
    },
...
  },
...
}